# api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/exam_service","description":"API Gateway"}],"paths":{"/api/questions/{id}":{"get":{"tags":["question-controller"],"operationId":"getQuestionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Question"}}}}}},"put":{"tags":["question-controller"],"operationId":"updateQuestion","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Question"}}}}}},"delete":{"tags":["question-controller"],"operationId":"deleteQuestion","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/exams/{id}":{"get":{"tags":["exam-controller"],"operationId":"getExamById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Exam"}}}}}},"put":{"tags":["exam-controller"],"operationId":"updateExam","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Exam"}}}}}},"delete":{"tags":["exam-controller"],"operationId":"deleteExam","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/questions":{"get":{"tags":["question-controller"],"operationId":"getAllQuestions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Question"}}}}}}},"post":{"tags":["question-controller"],"operationId":"createQuestion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Question"}}}}}}},"/api/exams":{"get":{"tags":["exam-controller"],"operationId":"getAllExams","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Exam"}}}}}}},"post":{"tags":["exam-controller"],"operationId":"createExam","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Exam"}}}}}}}},"components":{"schemas":{"Question":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"questionText":{"type":"string"},"optionA":{"type":"string"},"optionB":{"type":"string"},"optionC":{"type":"string"},"optionD":{"type":"string"},"correctAnswer":{"type":"string"},"marks":{"type":"integer","format":"int64"}}},"Exam":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"date":{"type":"string","format":"date"},"durationMinutes":{"type":"integer","format":"int64"},"status":{"type":"boolean"},"totalMarks":{"type":"integer","format":"int64"}}}}}}